Share authentication between reader and civetstorer
